L02

+++ CalcCLI/src/it/unibs/pajc/BinaryOperator.java
@@ -1,6 +1,6 @@
 package it.unibs.pajc;
 
 public interface BinaryOperator {
-	int eval(int a, int b);
+	double eval(double a, double b);
 
 }
